// project/src/app/interfaces/product.ts
export interface ProductInterface {
    id:          string
    name:        string
    price:       string
    image:       string
    active:      boolean
    seller:      string
    brand:       string
    model:       string
    color:       string
    category:    string
    description: string
}